+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - [asSet() for array and strings (#1)](https://github.com/ngarbezza/oow/issues/1)
+- [occurrencesOf(object) for array and strings (#2)](https://github.com/ngarbezza/oow/issues/2)
 
 ## [1.5.1] - 2019-02-23